refactor generate_test_data.py; update whats-new.rst
