+++ test_data/generate/convert_final_prms_output_to_nc.py
@@ -1,0 +1,40 @@
+from warnings import warn
+
+import pytest
+from prms_diagnostic_variables import (
+    diagnose_final_vars_to_nc,
+)
+
+import pywatershed as pws
+
+# We separate these files from the other convert_prms_output_to_nc.py files
+# because those need to be completely finished before these are run.
+
+# Also, dont run these tests in parallel.
+
+
+@pytest.fixture(scope="session")
+def final_netcdf_file(control_final_file):
+    """Create NetCDF files that depend on multiple other NetCDFs"""
+
+    control_file = control_final_file[0]
+    final_file = control_final_file[1]
+    domain_dir = final_file.parent
+    var_name = final_file.name
+    control = pws.Control.load_prms(control_file, warn_unused_options=False)
+    output_dir = control_file.parent / control.options["netcdf_output_dir"]
+
+    success = diagnose_final_vars_to_nc(
+        var_name, output_dir, control_file, domain_dir
+    )
+
+    if not success:
+        warn(
+            "make_final_netcdf_files False PASS above: "
+            f"unable to diagnose {final_file}"
+        )
+
+
+@pytest.mark.xdist_group(name="finalgroup")
+def make_final_netcdf_files(final_netcdf_file):
+    print(f"Creating final NetCDF file {final_netcdf_file}")
