+++ test_data/generate/convert_prms_output_to_nc.py
@@ -1,10 +1,6 @@
-from warnings import warn
-
 import pytest
-from filelock import FileLock
 from prms_convert_to_netcdf import convert_csv_to_nc, convert_soltab_to_nc
 from prms_diagnostic_variables import (
-    diagnose_final_vars_to_nc,
     diagnose_simple_vars_to_nc,
 )
 
@@ -12,7 +8,7 @@ import pywatershed as pws
 
 
 @pytest.fixture
-def netcdf_file(control_csv_file, tmp_path_factory):
+def netcdf_file(control_csv_file):
     """Convert CSV files from model output to NetCDF"""
     control_file = control_csv_file[0]
     csv_file = control_csv_file[1]
@@ -21,14 +17,9 @@ def netcdf_file(control_csv_file, tmp_path_factory):
     data_dir = csv_file.parent
     convert_csv_to_nc(var_name, data_dir)
 
-    root_tmpdir = tmp_path_factory.getbasetemp().parent
-    with FileLock(root_tmpdir / f"{var_name}.lock"):
-        yield csv_file  # do this in session cleanup
-        success = diagnose_simple_vars_to_nc(var_name, data_dir, control_file)
+    success = diagnose_simple_vars_to_nc(var_name, data_dir, control_file)
 
     if not success:
-        # should this fail or not?
-        # pytest.skip(f"Unable to diagnose {var_name}")
         assert False, "Unable to diagnose {var_name}"
 
     return f"{var_name}.nc"
@@ -52,35 +43,3 @@ def soltab_netcdf_file(tmp_path_factory, control_soltab_file):
 
 def make_soltab_netcdf_files(soltab_netcdf_file):
     print(f"Creating NetCDF files for soltab file {soltab_netcdf_file}")
-
-
-@pytest.fixture(scope="session")
-def final_netcdf_file(tmp_path_factory, control_final_file):
-    """Create NetCDF files that depend on multiple other NetCDFs"""
-
-    control_file = control_final_file[0]
-    final_file = control_final_file[1]
-    domain_dir = final_file.parent
-    var_name = final_file.name
-    control = pws.Control.load_prms(control_file, warn_unused_options=False)
-    output_dir = control_file.parent / control.options["netcdf_output_dir"]
-
-    root_tmpdir = tmp_path_factory.getbasetemp().parent
-    with FileLock(root_tmpdir / "final_nc.lock"):
-        yield final_file  # do this in session cleanup
-
-        success = diagnose_final_vars_to_nc(
-            var_name, output_dir, control_file, domain_dir
-        )
-
-        if not success:
-            warn(
-                "make_final_netcdf_files False PASS above: "
-                f"unable to diagnose {final_file}"
-            )
-
-    return final_file  # dosent really matter
-
-
-def make_final_netcdf_files(final_netcdf_file):
-    print(f"Creating final NetCDF file {final_netcdf_file}")
